src/bot/requirements.txt
functions-framework>=3.9.2
google-cloud-logging>=3.8.0
google-cloud-secret-manager>=2.20.0
python-box>=7.3.2
flask>=3.1.1
slack-bolt>=1.23.0